Add docs input `CheckboxList::class`.
